backend/pyproject.toml
[tool.poetry]
name = "market_data_api"
version = "0.1.0"
description = "get stocks and options related information"
authors = ["munikumar <munikumar7085@gmail.com>"]
license = "N/A"
readme = "README.md"
packages = [{include = "app"}]


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
requests = "^2.31.0"
smartapi-python = "^1.3.5"
websocket-client = "^1.5.2"
pyotp = "^2.9.0"
logzero = "^1.7.0"
httpx = "^0.25.0"
pytest = "^8.2"
dvc = {extras = ["gdrive"], version = "^3.49.0"}
platformdirs = "3.10.0"
numpy = "^1.26.4"
pandas = "^2.2.2"
sqlmodel = "^0.0.19"
registrable = "^0.0.4"
kafka-python-ng = "^2.2.2"
pytest-mock = "^3.14.0"
autobahn = "^24.4.2"
twisted = "^24.7.0"
python-dotenv = "^1.0.1"
brevo-python = "^1.1.2"
psycopg2-binary = "^2.9.10"
pyjwt = "^2.10.1"
bcrypt = "^4.2.1"
snowflake-id = "^1.0.2"
email-validator = "^2.2.0"
ruamel-yaml = "^0.18.9"
shfmt-py = "^3.7.0.1"
python-dateutil = "^2.9.0.post0"
pytest-asyncio = "^0.25.3"
fastapi-limiter = "^0.1.6"
fakeredis = {extras = ["lua"], version = "^2.26.2"}
playwright = "^1.50.0"



[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^24.4.2"
mypy = "^1.5.1"
pylint = "^3.0.1"
pytest-cov = "^5.0.0"
types-protobuf = "^5.29.1.20250208"
types-python-dateutil = "^2.9.0.20241206"
types-requests = "^2.32.0.20241016"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
